feat: Optimize device settings update and logging
This update enhances the `update_device_settings`, `lock_setting`, and `unlock_setting` functions by implementing prefetching of current device settings to reduce N+1 query issues. Additionally, it improves logging of setting changes by capturing old values before updates, locks, and unlocks, ensuring better historical fidelity. The `SettingsHistoryResponse` model is updated to reflect a change in the total count type from `i32` to `i64`, accommodating larger datasets. These changes improve performance and data integrity in device settings management.

+++ crates/domain/src/models/setting_change.rs
@@ -29,7 +29,7 @@ pub struct SettingsHistoryResponse {
     /// List of setting changes.
     pub changes: Vec<SettingChangeResponse>,
     /// Total number of changes for this device.
-    pub total_count: i32,
+    pub total_count: i64,
     /// Whether there are more results available.
     pub has_more: bool,
 }
